feat: Orthogonal Camera
Adds an orthogonal-perspective camera to the game.

To set it up, I created a Node3D containing a Camera3D node. I set the Camera3D perspective to `orthogonal`. This removes the FOV effect of the regular cameras, achieving the design level we want.

Next, I move the Camera position respect the Camera Node3D, and rotate the Node3D, which rotates the camera along with it. By setting the rotation to 45° in the respective axis, I achieve the camera effect I want.

Finally, I adjust the camera size to set the zoom to a level I consider adequate.