feat: Directional Light
In order to provide some sort of illumination to the scene I added a directional light to it. This light provides light for the whole scene, as if it was the sun.

I moved the node a bit aside from the scene for better visibility, but doing this transformation has no effect on the lighting.

To display shadows I had to enable the Shadows property in the inspector. I changed the mode to Orthogonal and adjusted the max distance for displaying shadows properly. Those properties are aligned in coordination with the camera `far` distance until I achieve the desired result.