fix: Ball physics
I wanted to add some bounciness to the ball. This way when it falls we can observe some bounciness to it.

I noticed the fall was not bouncing on the floor, but trespassing it all along. CSG boxes already set collision shapes, so I was not sure what was going on there.

The issue seemed to be related to the ball size, so small it was hard for the engine to handle.

I resized the ball and voilá! The collisions works, and I could even notice the shadows were working fine now.