feat: Mouse Raycaster
To hit the ball, we want to use a swipe mechanic where the user drags the finger on the screen to aim and shoot.

For this to work as expected, we need to transform the viewport coordinates of the player gestures into coordinates of the 3D world. For that, I implemented this script.

The script uses the given camera, object reference, and plane to translate viewport 2D coordinates to space 3D coordinates.

We use `project_ray_origin` to project the viewport 2D coordinates in a 3D space corresponding to the 2D dot in the 3D plane represented by the orthogonal camera. We combine this point with a direction vector parallel to the camera direction, and get the intersection of this vector with an horizontal plane set at the ball height.