feat: Apply impulse to ball
After refactoring the HUD and getting the vector to be displayed properly, it is time to implement the actual game mechanic of hitting the ball.

With the force and direction vector set, this is easy to do. Just apply an impulse with the vector direction and strength to the rigid body and let it go.

There is still some need to set the friction right to get the ball to slow down while sliding through the level.