feat: Adjust physics
I want the ball to emulate realistic physics, slowing down due to friction.

It seems, the physics engine does not handle well rotation due to the small contact surface of the ball with the floor. Thus, I had to tweak the `angular` `damp` parameter of the ball rigid body to simulate this friction effect.

There is still a lot to do in order to achieve a realistic simulation of the ball moving, but for now it should be enough. By playing with the angular damp and the force applied to the ball it is possible to get a more realistic effect.