chore: isometric camera independent from player node
Move the Camera node outside of the Player node. This is made due to the fact that the camera should not really depend on the player node. We want to be able to freely move the camera around in the future, so by moving it out of the node now we are preparing for the near future.

This also fixes an issue with the manual offset of the HUD by setting the position of the plane correctly when initializing the MouseRaycaster. Now we do not need additional adjustments in the vector to make it work.