chore: Level scene logic
Defines the logic of a level. A level is composed of a map, a starting position for the player ball and a finish area (where the ball should stay to consider the level completed).

The node uses a timer to count the amount of time the ball stays in the hole, confirming it stays there and does not leave before sending the win state.

The game scene will use this information to determine whether the player object is in the hole and the game is over.