progress: Level Editor
Refactors part of the Level Resource and Level Editor logic to reduce coupling.

Now the editor is working. I am able to dump my progress in the editor when running the LevelEditor scene, and loading it when I load a level resource.

There is still some issue to address. I am storing the meshes, but I am missing certain information, such as the rotation, which makes not possible to restore 100% accurately the map.