feat: update Level scene logic to handle LevelResource resources
Now the `Level` scene can handle `LevelResource`s. When the `Level``level_resource` attribute is set, it triggers a similar logic of the `LevelEditor`, loading the map and updating the level start point.

There is still some work to do in terms of adding a way to dumping the end area of the levels in the `LevelEditor`, and then load it in the `Level` scene, but for now we can already load different levels in a `Level` scene.