chore: `LevelSelector` Levels positioning in 3D space
Extended original `LevelSelector` logic for instantiating `Level` scenes for previsualization.

Now each level is rendered in a different position, so they do not overlap with each other. Level instances are wrapped inside a `Carrousel` node, whose purpose is to be manipulated in order to display the different levels in front of the camera.