feat: LevelNavigator
Adds a new `LevelNavigator` scene. This scene is responsible for letting the player navigate through the different levels in the game. It is a scene that, used in combination with the `LevelSelector`, allows players to previsualize the different levels, select one and start playing.

There is still the need to implement this level selection -> gameplay transition logic, which will come up in the next PR.