feat: LevelManager
Adds a new `LevelManager` class. This class is responsible of loading all game levels in memory, from their respective resources.

Now the `LevelNavigator` uses the `LevelManager` to access to the levels available in game, and navigate between those.

The PR also refactors parts of the existing logic to simplify the current logic