feat: Level world environment
Adds a world environment to the level, consisting in a sky panorama composed by a 2D Gradient texture which emulates sky and night.

Now we are progressing faster in the game, I want to start getting familiar with some of the more artistic parts of the game, and start giving it some look and feel. The world environment and lighting still looks awful, but at least I now know how to make gradient skies!