feat: Test Library
Creates a new mesh library for test levels.

This new mesh library includes better geometries and measures to help designing more rich levels. The other first iteration used bigger blocks with more restrictions. Now we are using smaller ones.

On top of that, we are using a world environment and directional light composition in the library editor to previsualize how a level can look like in the final game. We will reuse some of these composition parameters to define the level environment and lighting properties, aiming to create the feeling of different worlds by tunning those visual parameters.