feat: new test mesh library for levels
Implemented a new mesh library for designing game levels. This new mesh is way smaller than the original one, and has different colors assigned that blend better with the overall look and feel of the system.

However, when testing the actual game I can feel the ball bounces a lot on the floor. Like, somehow, the surface was not flat at all. Which it is very annoying.

I also need to design some smaller tiles for the walls, and a bigger tile for the hole.

For now it is fine, we are getting already a way better good-looking game as before, and we introduced some ideas on how to handle materials, lighting and so on. We should iterate further until we click with the right design.